+++ pytorch_lightning/accelerators/accelerator.py
@@ -41,6 +41,7 @@ class Accelerator:
     - CPU
     - GPU
     - TPU
+    - IPU
 
     Each Accelerator gets two plugins upon initialization:
     One to handle differences from the training routine and one to handle different precisions.
@@ -446,6 +447,17 @@ class Accelerator:
             If true, restore checkpoint after pre_dispatch.
         """
         return self.training_type_plugin.restore_checkpoint_after_pre_dispatch
+
+    def get_device_stats(self, device: Union[str, torch.device]) -> Dict[str, Any]:
+        """Gets stats for a given device.
+
+        Args:
+            device: device for which to get stats
+
+        Returns:
+            Dictionary of device stats
+        """
+        raise NotImplementedError
 
     def on_train_start(self) -> None:
         """Called when train begins."""